telegram_chess_bot/requirements.txt
chess
cairosvg
python-telegram-bot